+++ src/components/tree/tangle/untangling.js
@@ -79,6 +79,6 @@ export const untangleTreeToo = (phylotree1, phylotree2) => {
   flipChildrenPostorder(phylotree1, phylotree2);
   // console.log(`Untangling ${init_corr} -> ${calculatePearsonCorrelationCoefficient(phylotree1, phylotree2)}`);
   // TODO: check the value of streams
-  setDisplayOrder({nodes: phylotree2.nodes, focus: false, streams: false});
+  setDisplayOrder({nodes: phylotree2.nodes, focus: null, streams: false});
   // console.timeEnd("untangle");
 };